Add method InitializeRetailButtons() - Update method UpdateRetailButtons()
The front end UI binds with the observable collection RetailButtons. Everything seemed to work except for the visibility, which although does have it's own custom converter, was not working. All buttons were visible even though their default value is false for visibility. Thus, the collection must contain empty RetailButton objects so that it can bind to the UI Visibility property. This can possibly be refactored at a later date - and it probably had something to do with the visibility converter