Update method WebsiteClick() to use RetailButtonManager object
Additionally, the current method WebsiteClick() does not support different buttons yet. This should probably be in a different refactoring a little down the road. Something that could easily be done is changing the name of the buttons in the UI, for instance NonfoodButton1, and the substring Nonfood would tell the WebsiteClick() method in some kind of switch statement or if statement to use the Nonfood percentage